updates to include dockercoins,rollbacks, etc

+++ workshop/setup.d/install-stern.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
-mkdir -p /home/eduk8s/bin
-curl -L -o /home/eduk8s/bin/stern \
-   https://github.com/wercker/stern/releases/download/1.11.0/stern_linux_amd64
-chmod +x /home/eduk8s/bin/stern
+
+if [[ ! -e /home/eduk8s/bin/stern ]]; then
+  mkdir -p /home/eduk8s/bin
+  curl -L -o /home/eduk8s/bin/stern \
+    https://github.com/wercker/stern/releases/download/1.11.0/stern_linux_amd64
+  chmod +x /home/eduk8s/bin/stern
+fi
